Add permanent download url for artifacts to the output
Introduced a new output `BITRISE_PERMANENT_DOWNLOAD_URL_MAP` to hold
permanent urls for the artifacts uploaded by the step. Introduced a
new input `permanent_download_url_map_format` to control the format
of the output.

We received a request that it would be helpful if this step made
permanent download links available on its output for other steps to use.
Engagement team is responsible for generating these links. [EN-492]

[STEP-244]