[STEP-244] Permanent Download URLs (#127)
* Add permanent download url for artifacts to the output

Introduced a new output `BITRISE_PERMANENT_DOWNLOAD_URL_MAP` to hold
permanent urls for the artifacts uploaded by the step. Introduced a
new input `permanent_download_url_map_format` to control the format
of the output.

We received a request that it would be helpful if this step made
permanent download links available on its output for other steps to use.
Engagement team is responsible for generating these links. [EN-492]

[STEP-244]

* Fix step.yml formatting

* prevent golint from failing

* Add permanent urls even if public is disabled

* add logging for the endpoint

* Trigger PR build

* Remove debug logging

* Skip `*.aab` and `*.xcarchive.zip` from the public page

* Remove warning when artifact is not added to public page
* Add clarifying description to the public map output

* Change input and output descriptions to have clearer message

* Add check to the CI so it is ensured the step has the required output

* Avoid anti-pattern error print