[BINS-319] junit report generation fix when report includes a flaky rerun
Fixing the issue when the xcresult3 report includes a flaky test,
performed in "retry on failure" mode,
where the exact same test case first fails,
and then succeeds on the retry/rerun.

The previous implementation generated a junit report
which included both runs of the test case,
but both were marked as failed,
while the expected output would be a junit
report with a failed and then a successful run for the same (flaky) test case.